Better camera support.
Fix some framing issues, especially with UsdRender. Add depth of field and
shutter. General cleanup of the render pass logic to be less of an ugly mess.
Removed the auto environment light. Rewrite of the headlight to be actually
parented to the camera transform. Probably other stuff I forgot about.