Fix handling of disabled materials.
Hydra has a built-in way of doing this which works by returning an empty
material resource. This was not handled by our "default surface" code
which catches only a lack of material assignment. This commit fixes that
by connecting the default surface network when the material resource is
empty.

It also improves the update logic to not reexport networks which have
not actually changed.