Fixes to build with USD up to 22.08
Changes all relate to HdCamera which stopped storing view and projection
matrices between 21.11 and 22.03. Thankfully, we no longer need the code
from 709c29fb22 so it can be discarded instead of updated.

We still store our own copy of the projection matrix to avoid repeated
calls to ComputeProjectionMatrix(). Create() calls were moved around to
ensure this is populated before we call IsPerspective(). The rest is
just API name changes.