Render with a persistent NSI camera node.
The data from the selected camera from the scene is now copied onto our
rendering camera which uses a fixed handle and the same NSI node, as
long as its type does not change. This allows changing the source camera
without breaking down the output chain, which was forcing rendering to
stop.

This commit also enables efficient update of the aperture, exported as
screen window on the screen node. Previously, updating the screen always
involved stopping the render.

The reason for all this is Houdini which, when the viewport is locked to
a camera, temporarily changes from that camera to its "free viewport
cam" when it is dragged around. Then goes back to the correct camera
when the mouse is released. Stopping the render was making this quite
slow to update.