Export acceleration blur as multiple samples.
I had to use a scene index filter because husk does "something" with the data
otherwise. It somehow manufactures 3 time samples of points, velocities and
accelerations. And none of the values make any sense, making it impossible to
do the correct math in the hydra exporter. The filter manages to override that
behavior.